Enlit_github_repo/requirements.txt
pandas
scikit-learn
PyPDF2
python-docx
smtplib
